Add explicit (double) casts for float arithmetic under SSE math
With x87 FPU, float intermediates implicitly compute at 80-bit
extended precision. With -mfpmath=sse, float*float stays 32-bit.
Add explicit (double) casts where intermediate precision matters:

- util.cpp: sqrt() in VecToAngles
- dll.cpp: fabs() in door travel distance calculation
- tests: sqrt/sqrtf in vec_length, trace normal, and PVS distance

+++ tests/engine_mock.cpp
@@ -202,7 +202,7 @@ static edict_t *mock_pfnFindEntityInSphere(edict_t *pentStart,
       float dx = e->v.origin[0] - origin[0];
       float dy = e->v.origin[1] - origin[1];
       float dz = e->v.origin[2] - origin[2];
-      float dist = sqrtf(dx*dx + dy*dy + dz*dz);
+      float dist = sqrt((double)dx*dx + (double)dy*dy + (double)dz*dz);
 
       if (dist <= radius)
          return e;
